🔨 [templates] Move function `bindvariables` to `domain.binders`
